requirements.txt

asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.0.0
pip==24.0
python-decouple==3.8
sqlparse==0.5.3
mysqlclient
djangorestframework-simplejwt
gunicorn==23.0.0
